First GUI exists!
Recent development updates:
UI is in!
The barebones functional UI is in. You can officially train a neural net
from just the UI.
There are plenty of obvious things that need to be done to improve
it/fix it, but it technically works.

There is a lot to be done here, and there is a lot of low hanging fruit.
I encourage you to try and pick some
if the mood strikes you.

Text is OUT.
There is no more Text class and no trace of it in the project.

RNN Was Rewritten
When refactoring TrainingData, it was obvious that the RNN was basically
trying to learn a sequence of random numbers.
I am now 80% sure it is not doing that, and it has been rewritten to
work with the new TrainingData.
Test results are unaffected...

DLL Breaks
- TrainingData has been reworked.
Old TrainingData didn't support sequences, and was asymmetric in how it
handled data vs "label" (renamed "response")
New training data is clear and consistent,
Allows for sequencing and subsequencing.
Handles the different use cases in an intuitive way.

- Abstract Nets and Layers
Classes ANet and ALayer were created.
These offload some of the common work/definitions that exist in multiple
classes.
They have replaced the interface ILayer serve as INet.

Future development should derive from these classes.
More things will be factored out of them.

- The layer/net classes have been renamed/refactored/reorganized.
The unreadable mess that was these classes is now more like an
unreadable collection of evidently like objects.